+++ inst/extdata/non_documented_fun.R
@@ -1,0 +1,4 @@
+#' Non-documented function.
+#'
+#' @export
+fx <- function(x) x