+++ NEWS.md
@@ -1,6 +1,10 @@
-# exampletestr 0.3.0
-* First CRAN-worthy version.
+# 0.4.0
+* The package now uses .Rd files to read the examples rather than roxygen2 tags
+* Comments in examples are handled better
+* There is now the option to not put stuff in expect_equal() statements by default
 
-# exampletestr 0.3.1
+# 0.3.1
 * Minor fix to vignette.
 
+# exampletestr 0.3.0
+* First CRAN-worthy version.